+++ src/layers/MultiscaleImageLayer/utils.js
@@ -1,5 +1,5 @@
 import XRLayer from '../XRLayer';
-import ArrayBitmapLayer from '../ArrayBitmapLayer';
+import BitmapLayer from '../BitmapLayer';
 
 export function range(len) {
   return [...Array(len).keys()];
@@ -28,14 +28,13 @@ export function renderSubLayers(props) {
     isLensOn,
     lensSelection,
     onClick,
-    loader
+    loader,
+    modelMatrix
   } = props;
   // Only render in positive coorinate system
   if ([left, bottom, right, top].some(v => v < 0) || !data) {
     return null;
   }
-  const Layer =
-    loader.isRgb && loader.isInterleaved ? ArrayBitmapLayer : XRLayer;
   const { height, width } = loader.getRasterSize({ z: 0 });
   // Tiles are exactly fitted to have height and width such that their bounds match that of the actual image (not some padded version).
   // Thus the right/bottom given by deck.gl are incorrect since they assume tiles are of uniform sizes, which is not the case for us.
@@ -45,11 +44,27 @@ export function renderSubLayers(props) {
     data.width < loader.tileSize ? width : right,
     top
   ];
-  const layer = new Layer(props, {
-    id: `tile-sub-layer-${bounds}-${id}`,
+  const sharedLayerProps = {
     bounds,
-    channelData: data,
+    id: `tile-sub-layer-${bounds}-${id}`,
+    tileId: { x, y, z },
+    onHover,
     pickable,
+    onClick,
+    modelMatrix,
+    opacity,
+    visible
+  };
+  const { isRgb, isInterleaved, photometricInterpretation } = loader;
+  if (isRgb && isInterleaved) {
+    return new BitmapLayer(props, {
+      image: data,
+      photometricInterpretation,
+      ...sharedLayerProps
+    });
+  }
+  return new XRLayer(props, {
+    channelData: data,
     // Uncomment to help debugging - shades the tile being hovered over.
     // autoHighlight: true,
     // highlightColor: [80, 80, 80, 50],
@@ -57,16 +72,11 @@ export function renderSubLayers(props) {
     sliderValues,
     colorValues,
     channelIsOn,
-    opacity,
-    visible,
     dtype,
     colormap,
-    onHover,
     unprojectLensBounds,
     isLensOn,
     lensSelection,
-    tileId: { x, y, z },
-    onClick
+    ...sharedLayerProps
   });
-  return layer;
 }
